updated MessageController.getById parameter type to long updated MessageControllerTest to use mocked repository

+++ src/main/java/com/example/message/service/message/MessageController.java
@@ -37,7 +37,7 @@ public class MessageController {
     @GET
     @Path("/{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    public MessageResponse getById(@PathParam int id) {
+    public MessageResponse getById(@PathParam long id) {
         return messageService.getById(id);
     }
